TODO: fix crash if an excaption is thrown in a callback
Put a todo in the code to fix a bug. This is just so
it appears in the blame history. If an exception
is thrown in a callback called from c then an assert
will fail and the application will crash with a SIGABRT.

The real solution is to handle this gracefully in the c code
but I couldn't get it to work last time I tried.

steps to reproduce

diff --git a/test/issues.js b/test/issues.js
index 735b171..4e77eae 100644
--- a/test/issues.js
+++ b/test/issues.js
@@ -40,17 +41,20 @@ describe('Issues', function() {
             uri: uri,
             validatecert: LDAP.LDAP_OPT_X_TLS_NEVER,
             connect: function() {
+                throw new Error("oh no");

This is a workaround but i'm not sure if it's the
correct solution yet

diff --git a/main.js b/main.js
index a795965..9564d22 100644
--- a/main.js
+++ b/main.js
@@ -115,7 +115,11 @@ function LDAP(opt, fn) {

 LDAP.prototype.onconnect = function() {
   this.stats.reconnects++;
-  return this.options.connect.call(this);
+  try {
+    return this.options.connect.call(this);
+  } catch (e) {
+    console.error(e);
+  }
 };

It's worth noting that there are other callbacks called by c
that will also cause crashes.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>