fix test "Connect context should be ldap object - Issue #84"
I'm not sure where to start with this one but we have the following
code:

    it('Connect context should be ldap object - Issue #84', function(done) {
        ldap = new LDAP({
            uri: uri,
            validatecert: LDAP.LDAP_OPT_X_TLS_NEVER,
            connect: function() {
                assert(typeof this.bind === 'function');
                ldap.bind({binddn: 'cn=Manager,dc=sample,dc=com', password: 'secret'}, function(err) {
                    assert.ifError(err);
                    done();
                });
            }
        }, function (err) {
            assert.ifError(err);
        });
    });

It appears that the connect callback was being called instantly so
ldap wasn't actually being set and therefore
when we call ldap.bind we where instead using the ldap connection
from the previus test (the ldap variable is a global variable).

Interesting I would often but not all the time get the following
error: `LDAPError: Server is busy` from the test after the one
pasted above. I therefore skipped this test assuming it was
broken. Initially I found that replacing done() with
setTimeout(() => done(), 10) seemed to stop this from happening.
Later I found out why.

I'm not sure if this is a bug in the code and I've just
found a workaround or what, but it appears that the callback
you specify when you instantiate the LDAP class is called after
the bind event.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>