handle errors thrown in the connect callback
Previously if an error was thrown in the connect callback the
application would fail an assert in c and crash see bellow. This patch
handles the error in the c code and rethrows it.

node: ../cnx.c:1223: on_connect: Assertion `status == napi_ok' failed.
error Command failed with signal "SIGABRT".

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>