prevent the disconnect callback from being called twice
For whatever reason the disconnect callback was being called twice
when ldap.close was being called. As this functions stops the event
loop we therefore assume that if the even loop isn't running then
then there is no need to make a callback back to javascript as the
disconnect callback has already been called.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>