cnx: make all the other callbacks rethrow any caught exceptions
The issue was that if an exception was throw in a callback
called from c then an assert would fail and the application
would crash with a SIGABRT. This just rethrows the error in c
if it encounters one.

Unfortunately I don't know how to test these because I don't know
how I'm going to catch the error asynchronously.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>