Give the option of building openldap from source
Node 10 and above is statically linked to openssl 1.1.1 however
centos 7 uses openssl 1.0.2k and hense the version of opneldap
that comes with centos 7 is compiled against the old version of
openssl. This compiles openldap against the version of openssl
that Node10 and above uses and then statically link openldap
against the node module.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.com>