Commit libldap.a and statically link it
Node 10 and above is statically linked to openssl 1.1.1 however
centos 7 uses openssl 1.0.2k and hense the version of opneldap
that comes with centos 7 is compiled against the old version of
openssl. This contains a prebuild libldap.a and liblber.a built
against the version of openssl that Node10 and above ues.
Now if the major version of node is greater than 9 and your
using linux it'll statically link to these libraries.

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.com>