allow you to link the version of ldap install systemwide
By default it'll link to its own version of openldap. This means
you don't need openldap installed for this to work. It also ensures
that openldap is compiled with the same version of openssl that node
ueses.

You can link to the version of openldap installed systemwide by
specifying `USE_SYSTEM_LDAP=1 yarn`

Signed-off-by: Joshua Houghton <joshua.houghton@yandex.ru>