Fix use after free error
Segfault caused when ldap_abandon was being called after ldap_unbind
was called. This is due to this code here:

LDAP.prototype.enqueue = function(msgId, fn) {
  ...
  fn.timer = setTimeout(
    function searchTimeout() {
      /* in this instance the connection has already been closed */
      if (!this.ld) return;

      this.ld.abandon(msgid);
      delete this.queue[msgid];
      fn(new LDAPError("Timeout"));
      this.stats.timeouts++;
    }.bind(this),
    this.options.timeout
  );

For some this this.ld wasn't set to undefined which is odd because
it does get set to undefined here:

LDAP.prototype.close = function() {
  if (this.auth_connection !== undefined) {
    this.auth_connection.close();
  }
  this.ld.close();
  this.ld = undefined;
};

In short I'm not sure how this happens. The solution was to set
ldap_cnx->ld to NULL after unbinding it and then detecting if it's
NULL in the cnx_abandon function. I did throw an error when it's
NULL in the cnx_abandon function but this caused some test failures.

I may want to look into this some more in the future as I could very
easily be wrong in my assertions. It works for now anyway.

Steps to reproduce:

* Run the ./run_server script on a remote server and port forward
  port 1234 and 1235 to your local machine.
* Run the tests over and over again on your local machine and
  eventually you'll get a segfault.

Signed-off-by: Joshua Houghton <joshua.houghton@ripjar.com>