Fix a crash after socket disconnect/reconnect
Fixed a crash after the connection to the ldap server disconnects and
then reconnects.  This can happen connecting to some ldap servers as a
result of an idle timeout.  The crash is caused by reusing the old
handle for the new connection after already having called uv_poll stop.
Fixed by freeing the handle and setting the handle pointer to NULL.

Added a regression test for this case; it uses ldapjs to run an ldap
server that can be stopped/started as aprt of the test to simulate
being forcefully disconnected from the server.